Added meta.fix and meta.guard.
meta.guard!(t1, t2, ...) generates a template that tries instantiating the passed-in templates with its arguments and returns the result of the first succeeding template.  It's a try-catch-try-catch-... chain.

meta.fix is similar to meta.bind, but this one completely ignores its arguments and constantly evaluates to the bound template's result.  It would be useful for the fallback case of meta.guard.

TODO: Document them!