meta.uniqBy: Made ready for a scheduled change on reduceR.
I'll make a fix on meta.reduceR so that it returns "Seed" if a passed-in sequence is empty.  It will change reduceR to basically return an atomic entity, rather than sequence.

So, uniqBy must make it sure that the length of a sequence passed to reduceR should be larger than 1.